Working on the Sound Component
- Added the forms
- Already saves and loads
- Legacy project file support
- Needs to be implemented to the Builder
- Needs to be implemented to the Framework

+++ Source/JSGameIDE/JSGameIDE/Builder.cs
@@ -92,7 +92,8 @@ namespace JSGameIDE
             string TargetPath = customPath != null ? customPath : GameConfig.path + @"\Build\HTML5";
             try
             {
-                Directory.Delete(TargetPath, true);
+                if(Directory.Exists(TargetPath))
+                    Directory.Delete(TargetPath, true);
             }
             catch { }
             PreprocessorDefine();
